atara
Anxiety is extremely debilitating, and a common and extreme disturbance to ones peace one mind. So the absence of mental stress or anxiety(ataraxis) was born. Using HTML JS CSS - Through cross platform development this app will be available in all markets. An app that can be accessed anywhere at any time to help you remember yourself and how mind over matter is key-its not easy but together we can achieve "Ataraxis".